create productFactory.php
Add ProductsFactory for generating fake product data

This commit adds the ProductsFactory class to the database factories directory. The ProductsFactory is responsible for generating fake data for the Products model. It defines the default state for the model with attributes such as name, price, and quantity. This factory can be used to quickly generate dummy product data for development purposes.