Add font link to guest.blade file
This commit adds the font link to the HTML template, allowing the usage of the 'Figtree' font family with different weights. The font link is sourced from the URL 'https://fonts.bunny.net/css?family=figtree:400,500,600&display=swap'. The commit enhances the appearance of the website by utilizing the specified font family and weights.